% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMuts.R
\name{filterMuts}
\alias{filterMuts}
\title{Filter a GRangesList by Mutation Type}
\usage{
filterMuts(muts_grl, type = c("SNV", "DBS", "MBS", "INDEL"))
}
\arguments{
\item{muts_grl}{A \code{GRangesList} object, typically from
\code{importMuts}.}

\item{type}{A character string specifying the mutation type to
keep. One of "SNV", "DBS", "MBS", or "INDEL". (Case-insensitive)}
}
\value{
A \code{GRangesList} containing only mutations of the
specified type.
}
\description{
Subsets a \code{GRangesList} object to include only mutations of a
specified type (SNV, DBS, MBS, or INDEL).
}
\details{
This function is a wrapper for \code{MutationalPatterns::get_mut_type}.
It takes the full \code{GRangesList} from \code{importMuts} and
returns a new \code{GRangesList} containing only the desired
mutation type.

As per \code{MutationalPatterns}, DBS/MBSs are assumed to be
called as separate SNVs and will be merged into single variants.
}
\examples{
\dontrun{
  # Import all mutations
  muts_grl <- importMuts(vcf_files, genome = "hg19")

  # Filter for SNVs (for SBS analysis)
  snv_grl <- filterMuts(muts_grl, type = "SNV")

  # Filter for DBSs (for DBS analysis)
  dbs_grl <- filterMuts(muts_grl, type = "DBS")

  # Filter for INDELs (for ID analysis)
  indel_grl <- filterMuts(muts_grl, type = "INDEL")
}
}
\references{
Manders, F., Brandsma, A.M., de Kanter, J. et al. (2022).
MutationalPatterns: the one stop shop for the analysis of mutational
processes. \emph{BMC Genomics}, 23, 134.
\href{https://doi.org/10.1186/s12864-022-08357-3}{Link}
}
